fix test

+++ test/nc_stdlib_unittest.cpp
@@ -15,10 +15,10 @@ TEST(Stdlib, stackOrHeapAllocator) {
   StackOrHeapAllocator allocator(stack, 1024);
   EXPECT_EQ(allocator.allocArray<u8>(512) - stack, 0u);
   EXPECT_EQ(allocator.allocArray<u8>(512) - stack, 512u);
-  EXPECT_GT(allocator.allocArray<u8>(1) - stack, 4096u); // stack used up, so it's on heap
+  EXPECT_GT((size_t)(allocator.allocArray<u8>(1) - stack), 4096u); // stack used up, so it's on heap
 
   for (size_t i = 0; i < 100; i++) {
-    EXPECT_GT(allocator.allocArray<u8>(40960) - stack, 4096);
+    EXPECT_GT((size_t)(allocator.allocArray<u8>(40960) - stack), 4096u);
   }
 
   EXPECT_EQ(allocator._moreHeapPointerCount(), 93);
